+++ end_to_end_tests/collector_test.py
@@ -4,7 +4,6 @@ import os
 import shutil
 import tempfile
 import unittest
-from multiprocessing import Process
 
 import duckdb
 import pytest
@@ -16,21 +15,6 @@ from buildflow.dependencies.base import Scope, dependency
 from buildflow.io.local import File
 from buildflow.io.portable.table import AnalysisTable
 from buildflow.types.portable import FileFormat
-
-
-def run_flow(table: str):
-    app = buildflow.Flow()
-
-    @app.collector(
-        route="/test",
-        method="POST",
-        sink=AnalysisTable(table_name=table),
-        num_cpus=0.1,
-    )
-    def my_collector(input: InputRequest) -> OutputResponse:
-        return OutputResponse(input.val + 1)
-
-    app.run(start_runtime_server=True)
 
 
 @dataclasses.dataclass
@@ -73,40 +57,6 @@ class CollectorLocalTest(unittest.TestCase):
             os.remove(self.database)
         except FileNotFoundError:
             pass
-
-    def test_collector_duckdb_end_to_end_with_runtime_server(self):
-        try:
-            p = Process(target=run_flow, args=(self.table,))
-            p.start()
-
-            # wait for 20 seconds to let it spin up
-            self.get_async_result(asyncio.sleep(40))
-
-            response = requests.post(
-                "http://0.0.0.0:8000/test", json={"val": 1}, timeout=10
-            )
-            response.raise_for_status()
-
-            database = os.path.join(os.getcwd(), "buildflow_managed.duckdb")
-            conn = duckdb.connect(database=database, read_only=True)
-            got_data = conn.execute(f"SELECT count(*) FROM {self.table}").fetchone()
-
-            self.assertEqual(got_data[0], 1)
-
-            response = requests.get(
-                "http://127.0.0.1:9653/runtime/snapshot", timeout=10
-            )
-            response.raise_for_status()
-
-            self.assertEqual(response.json()["status"], "RUNNING")
-
-            response = requests.post("http://127.0.0.1:9653/runtime/drain", timeout=10)
-            response.raise_for_status()
-        finally:
-            p.join(timeout=20)
-            if p.is_alive():
-                p.kill()
-                p.join()
 
     def test_collector_file_end_to_end(self):
         output_dir = tempfile.mkdtemp()
